requirements.txt
rumps>=0.4.0
Pillow>=12.0.0
